First mapper implementation

+++ ric-k-means/src/main/java/it/unipi/cloud/hadoop/AggregateSamplesCombiner.java
@@ -1,8 +1,10 @@
 package it.unipi.cloud.hadoop;
 
 import it.unipi.cloud.model.PointWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class AggregateSamplesCombiner extends Reducer<LongWritable, PointWritable, LongWritable, PointWritable> {
+public class AggregateSamplesCombiner extends Reducer<IntWritable, PointWritable, IntWritable, PointWritable> {
+
 }
